added Middleware validator - UserController.update
