+++ src/routes.js
@@ -12,6 +12,7 @@ import NotificationController from './app/controllers/NotificationController';
 import AvailabeController from './app/controllers/AvailabeController';
 
 import validateUserStore from './app/validators/UserStore';
+import validateUserUpdate from './app/validators/UserUpdate';
 
 import authMiddleware from './app/middlewares/auth';
 
@@ -23,7 +24,7 @@ routes.post('/sessions', SessionController.store);
 
 routes.use(authMiddleware);
 
-routes.put('/users', UserController.update);
+routes.put('/users', validateUserUpdate, UserController.update);
 
 routes.get('/providers', ProviderController.index);
 routes.get('/providers/:providerId/available', AvailabeController.index);
